requirements.txt
requests==2.26.0
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
PyYAML==5.4.1
